+++ src/services/Handles.php
@@ -5,7 +5,6 @@ namespace kennethormandy\marketplace\services;
 use Craft;
 use craft\base\Component;
 use kennethormandy\marketplace\fields\MarketplaceConnectButton as MarketplaceConnectButtonField;
-use kennethormandy\marketplace\fields\MarketplacePayee as MarketplacePayeeField;
 use kennethormandy\marketplace\Marketplace;
 
 class Handles extends Component
@@ -26,29 +25,6 @@ class Handles extends Component
             foreach ($fields as $key => $value) {
                 if ($value instanceof MarketplaceConnectButtonField) {
                     return $value['handle'];
-                }
-            }
-        }
-
-        return null;
-    }
-
-    /**
-     * Get the custom field handle for the Marketplace Payee field.
-     * @deprecated The Payee field will be removed from the next major version of Marketplace, in favour of using events.
-     */
-    public function getPayeeHandle(): ?string
-    {
-        $usePayeeFieldType = Marketplace::$plugin->settings->usePayeeFieldType;
-        
-        if ($usePayeeFieldType) {
-            $fields = Craft::$app->fields->getAllFields();
-
-            if ($fields && count($fields) >= 1) {
-                foreach ($fields as $key => $value) {
-                    if ($value instanceof MarketplacePayeeField) {
-                        return $value['handle'];
-                    }
                 }
             }
         }